Validate model in service

+++ eight-ball/EightBall.Service/Services/BaseService.cs
@@ -1,6 +1,8 @@
-﻿using EightBall.Shared.Dtos;
+﻿using EightBall.Shared;
+using EightBall.Shared.Dtos;
 using EightBall.Shared.RepositoryInterfaces;
 using EightBall.Shared.ServiceInterfaces;
+using EightBall.Shared.Strings;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -25,29 +27,83 @@ namespace EightBall.Service.Services
             return _repository.EntityExists(id);
         }
 
-        public Task<TDto> GetByIdAsync(Guid id)
+        public async Task<Result<TDto>> GetByIdAsync(Guid id)
         {
-            return _repository.GetByIdAsync(id);
+            Result<TDto> result = new Result<TDto>();
+
+            var entityExists = await EntityExists(id);
+            if (!entityExists)
+            {
+                result.Errors.Add(Errors.NotFound, $"{typeof(TDto).Name} not found");
+
+                return result;
+            }
+
+            result.Succeeded = true;
+            result.Value = await _repository.GetByIdAsync(id);
+            return result;
         }
 
-        public Task<List<TDto>> GetEntitiesAsync()
+        public async Task<Result<List<TDto>>> GetEntitiesAsync()
         {
-            return _repository.GetEntitiesAsync();
+            Result<List<TDto>> result = new Result<List<TDto>>();
+            var tables = _repository.GetEntitiesAsync();
+
+            result.Succeeded = true;
+            result.Value = await tables;
+            return result;
         }
 
-        public Task<Guid> InsertAsync(TDto dto)
+        public virtual async Task<Result<Guid>> InsertAsync(TDto dto)
         {
-            return _repository.AddEntityAsync(dto);
+            Result<Guid> result = new Result<Guid>();
+            var inserted = await _repository.AddEntityAsync(dto);
+            if (inserted != Guid.Empty)
+            {
+                result.Succeeded = true;
+                result.Value = inserted;
+            }
+
+            return result;
         }
 
-        public Task<int> RemoveAsync(Guid id)
+        public async Task<Result> RemoveAsync(Guid id)
         {
-            return _repository.RemoveEntityAsync(id);
+            Result result = new Result();
+            var entityExists = await EntityExists(id);
+            if (!entityExists)
+            {
+                result.Errors.Add(Errors.NotFound, $"{typeof(TDto).Name} not found");
+                return result;
+            }
+
+            var deletedRows = await _repository.RemoveEntityAsync(id);
+            if (deletedRows != default)
+            {
+                result.Succeeded = true;
+            }
+
+            return result;
         }
 
-        public Task<int> UpdateAsync(TDto dto)
+        public virtual async Task<Result> UpdateAsync(TDto dto)
         {
-            return _repository.UpdateEntityAsync(dto);
+            Result result = new Result();
+            var entityExists = await EntityExists(dto.Id);
+            if (!entityExists)
+            {
+                result.Errors.Add(Errors.NotFound, $"{typeof(TDto).Name} not found");
+
+                return result;
+            }
+
+            int updatedRows = await _repository.UpdateEntityAsync(dto);
+            if (updatedRows != default)
+            {
+                result.Succeeded = true;
+            }
+
+            return result;
         }
     }
 }
